+++ src/haven_types.cpp
@@ -2,12 +2,24 @@
 #include <cmath>
 #include "haven_types.h"
 
-std::string formatAttribute(FileType type) {
-  switch (type) {
+FileVendor extVendor(FileExt ext) {
+  switch (ext) {
+  case HAVEN_DTA:      return HAVEN_STATA;
+  case HAVEN_SAV:
+  case HAVEN_POR:      return HAVEN_SPSS;
+  case HAVEN_SAS7BDAT:
+  case HAVEN_SAS7BCAT:
+  case HAVEN_XPT:      return HAVEN_SAS;
+  default:
+    Rcpp::stop("Unknown file extension");
+  }
+}
+
+std::string formatAttribute(FileVendor vendor) {
+  switch (vendor) {
   case HAVEN_STATA: return "format.stata";
   case HAVEN_SPSS:  return "format.spss";
   case HAVEN_SAS:   return "format.sas";
-  case HAVEN_XPT:   return "format.xpt";
   }
 
   return "";
@@ -29,14 +41,13 @@ VarType numType(SEXP x) {
   }
 }
 
-VarType numType(FileType type, const char* var_format) {
+VarType numType(FileVendor vendor, const char* var_format) {
   if (var_format == NULL)
     return HAVEN_DEFAULT;
 
   std::string format(var_format);
 
-  switch(type) {
-  case HAVEN_XPT:
+  switch(vendor) {
   case HAVEN_SAS:
     // http://support.sas.com/documentation/cdl/en/lrdict/64316/HTML/default/viewer.htm#a000589916.htm
     if      (hasPrefix(format,"DATETIME")) return HAVEN_DATETIME;
@@ -73,9 +84,8 @@ VarType numType(FileType type, const char* var_format) {
 
 // Value conversion -----------------------------------------------------------
 
-int daysOffset(FileType type) {
-  switch(type) {
-  case HAVEN_XPT:
+int daysOffset(FileVendor vendor) {
+  switch(vendor) {
   case HAVEN_SAS:   return 3653;   // 1960-01-01
   case HAVEN_STATA: return 3653;
   case HAVEN_SPSS:  return 141428; // 1582-01-01
@@ -84,19 +94,19 @@ int daysOffset(FileType type) {
   return 0;
 }
 
-double adjustDatetimeToR(FileType file, VarType var, double value) {
+double adjustDatetimeToR(FileVendor vendor, VarType var, double value) {
   if (std::isnan(value))
     return value;
 
-  double offset = daysOffset(file);
+  double offset = daysOffset(vendor);
 
   switch(var) {
   case HAVEN_DATETIME:
-    if (file == HAVEN_STATA) // stored in milliseconds
+    if (vendor == HAVEN_STATA) // stored in milliseconds
       value /= 1000;
     return value - offset * 86400;
   case HAVEN_DATE:
-    if (file == HAVEN_SPSS) // stored in seconds
+    if (vendor == HAVEN_SPSS) // stored in seconds
       value /= 86400;
     return value - offset;
   default:
@@ -104,21 +114,21 @@ double adjustDatetimeToR(FileType file, VarType var, double value) {
   }
 }
 
-double adjustDatetimeFromR(FileType file, SEXP col, double value) {
+double adjustDatetimeFromR(FileVendor vendor, SEXP col, double value) {
   if (std::isnan(value))
     return value;
 
-  double offset = daysOffset(file);
+  double offset = daysOffset(vendor);
 
   switch(numType(col)) {
   case HAVEN_DATETIME:
     value += offset * 86400;
-    if (file == HAVEN_STATA) // stored in milliseconds
+    if (vendor == HAVEN_STATA) // stored in milliseconds
       value *= 1000;
     return value;
   case HAVEN_DATE:
     value += offset;
-    if (file == HAVEN_SPSS) // stored in seconds
+    if (vendor == HAVEN_SPSS) // stored in seconds
       value *= 86400;
     return value;
   default: